Removed JQuery Dependency
There was a security scanner listed JQuery library as a security vulnerability and wanted it removed.  That, in addition to the fact that browsers natively support everything we needed from JQuery made it an obvious move (and reduced download requirement of the JQuery file).

Any conversions that were non-trivial, I use https://youmightnotneedjquery.com as a resource
- .parents(), .on(), .off()
- Application Helpers

Existing bugs found during development:
- @event/v-on handlers were lost on helptips/modals if assigned before cloning.
- Helptips we closing prematurely for `-textbox` templates - they were wrapped in `<label/>` so the help tip was created inside label and triggered multiple clicks (see comment inside `getTipContainer`)
- contentSelector modal apps losing registered events (didn't use `.clone(true)`)
- `hideLoader` call only hooked to `calculation` event now (before was also on configureUICalculation so ran code twice)